+++ src/dummy_compressor.cpp
@@ -5,8 +5,8 @@
 void DummyCompressor::compress(obstream& obs, vector<GPSPoint> points)
 {
   obs.write_int(points.size(), 32);
-  
-  for(int i = 0; i < points.size(); ++i)
+
+  for(size_t i = 0; i < points.size(); ++i)
   {
     obs.write_double(points[i].get_time());
     obs.write_double(points[i].get_latitude());
@@ -18,9 +18,9 @@ vector<GPSPoint> DummyCompressor::decompress(ibstream& ibs)
 {
   int size = ibs.read_int(32);
   vector<GPSPoint> points;
-  
+
   bits64 tuple[3];
-  
+
   for(int i = 0; i < size; ++i)
   {
     tuple[0].lng = ibs.read_int(64);
@@ -28,7 +28,7 @@ vector<GPSPoint> DummyCompressor::decompress(ibstream& ibs)
     tuple[2].lng = ibs.read_int(64);
     points.push_back(GPSPoint(tuple[0].dbl, tuple[1].dbl, tuple[2].dbl));
   }
-         
+
   return points;
 }
